final run

+++ src/QueryOne.cpp
@@ -97,7 +97,7 @@ void QueryOne::op_sum_qty(const table& tab, const std::vector<BITMAP_TYPE>& bitm
     for (int i = 0; i < size; i++) {
         auto key = std::pair<char, char>(tab.l_returnflag[i], tab.l_linestatus[i]);
         auto& group = groups.at(key);
-        group.sum_qty = db_plus(group.sum_qty,db_times((int)bitmap[i], tab.l_quantity[i]));
+        group.sum_qty = db_plus(group.sum_qty,db_times(bitmap[i], tab.l_quantity[i]));
     }
 }
 
@@ -107,7 +107,7 @@ void QueryOne::op_sum_base_price(const table & tab, const std::vector<BITMAP_TYP
     for (int i = 0; i < size; i++) {
         auto key = std::pair<char, char>(tab.l_returnflag[i], tab.l_linestatus[i]);
         auto& group = groups.at(key);
-        group.sum_base_price = db_plus(group.sum_base_price,db_times((int)bitmap[i], tab.l_extendedprice[i]));
+        group.sum_base_price = db_plus(group.sum_base_price,db_times(bitmap[i], tab.l_extendedprice[i]));
     }
 }
 
@@ -117,7 +117,7 @@ void QueryOne::op_sum_disk_price(const table & tab, const std::vector<BITMAP_TYP
     for (int i = 0; i < size; i++) {
         auto key = std::pair<char, char>(tab.l_returnflag[i], tab.l_linestatus[i]);
         auto& group = groups.at(key);
-        group.sum_disc_price = db_plus(group.sum_disc_price,db_times((int)bitmap[i], db_times(tab.l_extendedprice[i], db_minus(100,tab.l_discount[i]))));
+        group.sum_disc_price = db_plus(group.sum_disc_price,db_times(bitmap[i], db_times(tab.l_extendedprice[i], db_minus(100,tab.l_discount[i]))));
     }
 }
 
@@ -128,7 +128,7 @@ void QueryOne::op_sum_charge(const table & tab, const std::vector<BITMAP_TYPE>& 
         auto key = std::pair<char, char>(tab.l_returnflag[i], tab.l_linestatus[i]);
         auto& group = groups.at(key);
         auto currentCharge = db_times(tab.l_extendedprice[i], double(db_times(db_minus(100,tab.l_discount[i]),db_minus(100,tab.l_tax[i])))  / 10000.0);
-        group.sum_charge = db_plus(group.sum_charge, db_times(currentCharge, (int)bitmap[i]));
+        group.sum_charge = db_plus(group.sum_charge, db_times(currentCharge, bitmap[i]));
     }
 }
 
@@ -152,7 +152,7 @@ void QueryOne::op_avg_disc(const table & tab, const std::vector<BITMAP_TYPE>& bi
     for (int i = 0; i < size; i++) {
         auto key = std::pair<char, char>(tab.l_returnflag[i], tab.l_linestatus[i]);
         auto& group = groups.at(key);
-        group.avg_disc = db_plus(group.avg_disc,db_times(tab.l_discount[i], (int)bitmap[i]));
+        group.avg_disc = db_plus(group.avg_disc,db_times(tab.l_discount[i], bitmap[i]));
     }
     for (auto& group : groups) {
         group.second.avg_disc = db_divided(group.second.avg_disc,double(group.second.count_order));
